feature: weight composition data
This is the raw code that was tailored to sablefish for weighting the
input sample size of composition data using either the McAllister or the
Francis method.

NOTE - this code is not currently generalized and was specific to the
sablefish stock assessment.

One needs to think about how to change the model if the DM parameters are
already included in the model and need to be turned off before weighting
using these iterative methods. May also need to include DM as a method where
it will remove the added weighting lines and add DM parameter lines.
Lots of things are possible with time :)